challenge-1/requirements.txt
torch>=1.10.0
torchvision>=0.11.0
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.3
seaborn>=0.11.2
Pillow>=8.4.0
opencv-python>=4.5.4
scikit-learn>=0.24.2
tqdm>=4.62.3
